spotify: make more consistent with youtube
