+++ src/router/mod.rs
@@ -5,8 +5,6 @@ use std::sync::Arc;
 use std::thread;
 use std::time::{Duration, Instant};
 
-use tokio::sync::mpsc;
-
 use crate::spotify;
 use crate::midi;
 use crate::youtube;
@@ -34,9 +32,8 @@ pub struct RunConfig {
 pub struct Router {
     config: RunConfig,
     term: Arc<AtomicBool>,
-    spotify_app: spotify::SpotifyTaskSpawner<LaunchpadProEvent>,
-    spotify_receiver: mpsc::Receiver<spotify::Out<LaunchpadProEvent>>,
     server: HttpServer,
+    spotify_app: spotify::app::Spotify<LaunchpadProEvent>,
     youtube_app: youtube::app::Youtube<LaunchpadProEvent>,
     selected_app: AppName,
 }
@@ -45,17 +42,15 @@ impl Router {
     pub fn new(config: RunConfig) -> Self {
         let term = Arc::new(AtomicBool::new(false));
 
-        let (sp_sender, sp_receiver) = mpsc::channel::<spotify::Out<LaunchpadProEvent>>(32);
-        let spotify_app = spotify::SpotifyTaskSpawner::new(config.spotify_app_config.clone(), sp_sender);
+        let spotify_app = spotify::app::Spotify::new(config.spotify_app_config.clone());
         let server = HttpServer::start();
         let youtube_app = youtube::app::Youtube::new(config.youtube_api_key.clone(), config.youtube_playlist_id.clone());
 
         return Router {
             config,
             term,
-            spotify_app,
-            spotify_receiver: sp_receiver,
             server,
+            spotify_app,
             youtube_app,
             selected_app: AppName::Spotify,
         };
@@ -97,13 +92,13 @@ impl Router {
                 let launchpad_result = match launchpad.as_mut() {
                     Ok(launchpad) => {
                         let _ = LaunchpadProEvent::from_app_colors(vec![
-                            spotify::COLOR,
+                            spotify::app::COLOR,
                             youtube::app::COLOR,
                         ]).and_then(|event| launchpad.write(event));
 
                         match self.selected_app {
                             AppName::Spotify => {
-                                let event = self.spotify_receiver.try_recv();
+                                let event = self.spotify_app.receive();
                                 match event {
                                     Ok(spotify::Out::Command(command)) => {
                                         let _ = self.server.send(command);
@@ -134,7 +129,7 @@ impl Router {
                                     Ok(Some(0)) => {
                                         println!("Selecting Spotify");
                                         self.selected_app = AppName::Spotify;
-                                        let _ = LaunchpadProEvent::from_image(spotify::get_spotify_logo())
+                                        let _ = LaunchpadProEvent::from_image(spotify::app::get_spotify_logo())
                                             .and_then(|event| launchpad.write(event));
                                     },
                                     Ok(Some(1)) => {
@@ -145,7 +140,7 @@ impl Router {
                                     },
                                     _ => {
                                         match self.selected_app {
-                                            AppName::Spotify => self.spotify_app.handle(event),
+                                            AppName::Spotify => self.spotify_app.send(event),
                                             AppName::Youtube => self.youtube_app.send(event),
                                         }
                                     },
